% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar}
\alias{calendar}
\title{Plot a calendar}
\usage{
calendar(.data, event)
}
\arguments{
\item{.data}{A data frame.}

\item{event}{A variable to calendar.}
}
\value{
A ggplot.
}
\description{
\code{calendar()} plots a yearly calendar.
}
