% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_frame.R
\name{calendar_frame}
\alias{calendar_frame}
\title{Create a calendar data frame}
\usage{
calendar_frame(.date_range)
}
\arguments{
\item{.date_range}{A range of dates.}
}
\value{
A nested tibble.
}
\description{
\code{calendar_frame()} takes a range of dates and creates a data frame with the
necessary variables for plotting a yearly calendar.
}
