% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_nested.R
\name{calendar_nested}
\alias{calendar_nested}
\title{Create nested, yearly calendars}
\usage{
calendar_nested(.data, nest_by, event)
}
\arguments{
\item{.data}{A data frame.}

\item{nest_by}{A grouping variable other than year.}

\item{event}{A variable to calendar.}
}
\value{
A nested tibble.
}
\description{
\code{calendar_nested()} nests data by year and group given by \code{x} and adds a
list of yearly calendars of \code{y}.
}
