% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_year_to_fiscal.R
\name{convert_year_to_fiscal}
\alias{convert_year_to_fiscal}
\title{Convert calendar year to fiscal}
\usage{
convert_year_to_fiscal(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\description{
Convert calendar year to fiscal
}
