% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employer-package.R
\docType{package}
\name{employer-package}
\alias{employer}
\alias{employer-package}
\title{employer: R package for Human Resource Transactions}
\description{
Contains internal data, reports, and code outside of Shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asadow/employer}
  \item Report bugs at \url{https://github.com/asadow/employer/issues}
}

}
\author{
\strong{Maintainer}: Adam Sadowski \email{asadowsk@uoguelph.ca}

}
\keyword{internal}
