% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_dt.R
\name{formatted_dt}
\alias{formatted_dt}
\title{Creates a PR-prototypical DT.}
\usage{
formatted_dt(.data, pane_false, visible_false, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{pane_false}{Columns to not have search panes.}

\item{visible_false}{Columns to be invisible in the table.}

\item{...}{More arguments to \code{DT::datatable(options = ...)}.}
}
\description{
\code{formatted_dt()} creates an HTML widget as per the \code{DT} package.
}
