% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidate_frame.R
\name{holidate_frame}
\alias{holidate_frame}
\title{Create tibble of dates and holidays}
\usage{
holidate_frame(date, holidates)
}
\arguments{
\item{date}{A vector of class \code{Date}.}

\item{holidates}{A data frame with a \code{date} column for holiday dates.}
}
\description{
\code{holidate_frame()} takes a vector of dates
and returns a tibble with every date and holiday inside its range.
}
