% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holidates}
\alias{holidates}
\title{University of Guelph holiday dates from 2014-2023}
\format{
\code{holidates}: A data frame with
870 rows,
and 2 columns:
\describe{
\item{holiday}{Holiday name.}
\item{date}{Date.}
}
}
\usage{
holidates
}
\description{
University of Guelph holiday dates from 2014-2023
}
\keyword{datasets}
