% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_consecutive_work_days.R
\name{id_consecutive_work_days}
\alias{id_consecutive_work_days}
\title{Create id column identifying sets of continuous working days}
\usage{
id_consecutive_work_days(.data, x, work_days = 2:6, holidates)
}
\arguments{
\item{.data}{A data frame with a \code{date} column of class \code{Date}.}

\item{x}{An event variable of interest.}

\item{work_days}{Numeric vector of working days of the week (1 is Sunday).}

\item{holidates}{A data frame with a \code{date} column for holiday dates.}
}
\value{
A data frame with a new \code{id} column of class \code{integer}.
}
\description{
Create id column identifying sets of continuous working days
}
