% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_matrix.R
\name{tr_matrix}
\alias{tr_matrix}
\title{Matricize with totals of both rows and columns}
\usage{
tr_matrix(.data, by)
}
\arguments{
\item{.data}{A data frame.}

\item{by}{A variable to summarise by in addition to \code{days}.}
}
\value{
A tibble.
}
\description{
\code{tr_matrix()} creates a matrix summary out of summarized transactions.
}
