% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_pivot_years_wider.R
\name{tr_pivot_years_wider}
\alias{tr_pivot_years_wider}
\title{Pivot wide transaction summaries by year}
\usage{
tr_pivot_years_wider(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\description{
Pivot wide transaction summaries by year
}
