% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_summarize_runs.R
\name{tr_summarize_runs}
\alias{tr_summarize_runs}
\title{Summarize transaction runs: how many days per run}
\usage{
tr_summarize_runs(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\value{
A tibble.
}
\description{
\code{tr_summarize_runs()} creates nested summaries by runs.
}
