% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_of_month.R
\name{week_of_month}
\alias{week_of_month}
\title{Get week of month where weeks start on Sunday after the first week}
\usage{
week_of_month(x)
}
\arguments{
\item{x}{A date.}
}
\value{
A numeric.
}
\description{
\code{week_of_month()} calculates the week of the month.
}
