% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_half.R
\name{year_half}
\alias{year_half}
\title{Get year-half}
\usage{
year_half(date)
}
\arguments{
\item{date}{A vector of class date.}
}
\value{
A factor vector.
}
\description{
\code{year_half()} gives the January to June or July to December year-half.
}
