% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employer-package.R
\docType{package}
\name{employer-package}
\alias{employer}
\alias{employer-package}
\title{employer: R package for Human Resources at the University of Guelph}
\description{
Contains internal data, reports, and code outside of Shiny applications.
}
\author{
\strong{Maintainer}: Adam Sadowski \email{asadowsk@uoguelph.ca}

}
\keyword{internal}
